+++ src/main/java/com/example/crud/repository/TodoRepository.java
@@ -3,4 +3,4 @@ package com.example.crud.repository;
 import com.example.crud.model.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoRepository extends JpaRepository<Todo, Long> { }
+public interface TodoRepository extends JpaRepository<Todo, Integer> { }
